Separate CLI code as new file in scripts directory
